requirements.txt
altair==5.1.2
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
Faker==20.0.3
favicon==0.7.0
fonttools==4.44.3
gitdb==4.0.11
GitPython==3.1.40
htbuilder==0.6.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
plotly-express==0.4.1
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.16.1
pymdown-extensions==10.4
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.0
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
st-annotated-text==4.0.1
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.28.2
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.5
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.10
zipp==3.17.0
